Migration completed
The project has been converted to maven. And the necessary adjustments have been made to the code.
