--- .idea/artifacts/registrationForm_jar.xml
@@ -1,10 +1,0 @@
-<component name="ArtifactManager">
-  <artifact type="jar" name="registrationForm:jar">
-    <output-path>$PROJECT_DIR$/out/artifacts/registrationForm_jar</output-path>
-    <root id="archive" name="registrationForm.jar">
-      <element id="module-output" name="registrationForm" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/out/artifacts/registrationForm_jar/registrationForm.jar" path-in-jar="/" />
-      <element id="extracted-dir" path="$PROJECT_DIR$/lib/flatlaf-3.5.1.jar" path-in-jar="/" />
-    </root>
-  </artifact>
-</component>
