--- .idea/libraries/flatlaf_3_5_1.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="flatlaf-3.5.1">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/lib/flatlaf-3.5.1.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
