--- .idea/libraries/registrationForm.xml
@@ -1,9 +1,0 @@
-<component name="libraryTable">
-  <library name="registrationForm">
-    <CLASSES>
-      <root url="jar://$PROJECT_DIR$/out/artifacts/registrationForm_jar/registrationForm.jar!/" />
-    </CLASSES>
-    <JAVADOC />
-    <SOURCES />
-  </library>
-</component>
