+++ src/main/java/META-INF/MANIFEST.MF
